Eliminate "tearing" when turning quickly.
The reason for this was that chunk visibility flags could be out of date
by one frame. These changes force them to be recalculated as soon as we
know the camera has moved. If rendering is occuring during a visibility
update then tearing may still occur, but it has been greatly reduced.