Fetch chunks nearest the camera first.
When the active region is large, it is necessary to produce chunks near
the camera before attempting to produce chunks far away. Also, it's not
necessary to produce them all immediately so long as the nearby ones show
up quickly.

The end result is that frame rates and subjective performance of terrain
loading/generation is much faster now when the active region extent is
large.