Prevent slow chunk visibility update by having the draw thread fetch VAOs.
This gives the lowest latency between fetching and drawing VAOs. In order
to prevent the fetching from tanking frame rate, do it non-blocking. In
order to prevent the non-blocking fetch from having holes in it, don't let
go of chunks, once acquired, until they leave the camera frustum.

When placing or removing blocks, take steps to ensure the update appears
"instantly" on the draw thread.