Taking steps to avoid flcikers when placing blocks.
The active region draw method tries to avoid leaving a blank space in the
draw list. It keeps old chunks around if the new one is not ready. This
prevents flicker when making edits.

Also, added a journal for chunk edits, and the ability to reconstruct
chunks from the journal when they fail validation after the chunk loads.