First pass attempt at restructuring locking in GSGrid.
Now, you retrieve a Slot from a Grid, not the item itself. The slot can
be locked for reading or writing and the slot holds the item. This allows
individual chunk positions to be locked in the grid without also causing
the buckets array to be locked or the bucket to be locked.

Grid cost limits have been temporarily removed until I figure out a good
way to reimplement them.